+++ Core/Driver/DriverLoading/BindPCI.cpp
@@ -133,7 +133,7 @@ namespace Driver
 
         if (fexExtended->Driver.OverrideOnConflict)
         {
-            std::vector<int> DriversToRemove = std::vector<int>();
+            std::vector<uint64_t> DriversToRemove = std::vector<uint64_t>();
             foreach (auto Drv in Drivers)
             {
                 FexExtended *fe = ((FexExtended *)((uintptr_t)Drv.Address + EXTENDED_SECTION_ADDRESS));
@@ -181,7 +181,7 @@ namespace Driver
 
         if (fexExtended->Driver.OverrideOnConflict)
         {
-            std::vector<int> DriversToRemove = std::vector<int>();
+            std::vector<uint64_t> DriversToRemove = std::vector<uint64_t>();
             foreach (auto Drv in Drivers)
             {
                 FexExtended *fe = ((FexExtended *)((uintptr_t)Drv.Address + EXTENDED_SECTION_ADDRESS));
@@ -229,7 +229,7 @@ namespace Driver
 
         if (fexExtended->Driver.OverrideOnConflict)
         {
-            std::vector<int> DriversToRemove = std::vector<int>();
+            std::vector<uint64_t> DriversToRemove = std::vector<uint64_t>();
             foreach (auto Drv in Drivers)
             {
                 FexExtended *fe = ((FexExtended *)((uintptr_t)Drv.Address + EXTENDED_SECTION_ADDRESS));
@@ -306,7 +306,7 @@ namespace Driver
 
         if (fexExtended->Driver.OverrideOnConflict)
         {
-            std::vector<int> DriversToRemove = std::vector<int>();
+            std::vector<uint64_t> DriversToRemove = std::vector<uint64_t>();
             foreach (auto Drv in Drivers)
             {
                 FexExtended *fe = ((FexExtended *)((uintptr_t)Drv.Address + EXTENDED_SECTION_ADDRESS));
@@ -383,7 +383,7 @@ namespace Driver
 
         if (fexExtended->Driver.OverrideOnConflict)
         {
-            std::vector<int> DriversToRemove = std::vector<int>();
+            std::vector<uint64_t> DriversToRemove = std::vector<uint64_t>();
             foreach (auto Drv in Drivers)
             {
                 FexExtended *fe = ((FexExtended *)((uintptr_t)Drv.Address + EXTENDED_SECTION_ADDRESS));
@@ -431,7 +431,7 @@ namespace Driver
 
         if (fexExtended->Driver.OverrideOnConflict)
         {
-            std::vector<int> DriversToRemove = std::vector<int>();
+            std::vector<uint64_t> DriversToRemove = std::vector<uint64_t>();
             foreach (auto Drv in Drivers)
             {
                 FexExtended *fe = ((FexExtended *)((uintptr_t)Drv.Address + EXTENDED_SECTION_ADDRESS));
@@ -479,7 +479,7 @@ namespace Driver
 
         if (fexExtended->Driver.OverrideOnConflict)
         {
-            std::vector<int> DriversToRemove = std::vector<int>();
+            std::vector<uint64_t> DriversToRemove = std::vector<uint64_t>();
             foreach (auto Drv in Drivers)
             {
                 FexExtended *fe = ((FexExtended *)((uintptr_t)Drv.Address + EXTENDED_SECTION_ADDRESS));
